+++ backend/localrep/models/__init__.py
@@ -1,7 +1,8 @@
 from .algo import Algo
+from .computeplan import ComputePlan
 from .datamanager import DataManager
 from .datasample import DataSample
 from .event import Event
 from .metric import Metric
 
-__all__ = ["Algo", "DataManager", "DataSample", "Event", "Metric"]
+__all__ = ["Algo", "ComputePlan", "DataManager", "DataSample", "Event", "Metric"]